requirements.txt
fastapi
uvicorn
pandas
openai
python-dotenv
google-api-python-client 
google-auth
google_auth_oauthlib